+++ app/src/main/java/net/bible/android/view/activity/settings/ColorSettings.kt
@@ -1,0 +1,143 @@
+/*
+ * Copyright (c) 2020 Martin Denham, Tuomas Airaksinen and the And Bible contributors.
+ *
+ * This file is part of And Bible (http://github.com/AndBible/and-bible).
+ *
+ * And Bible is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later version.
+ *
+ * And Bible is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with And Bible.
+ * If not, see http://www.gnu.org/licenses/.
+ *
+ */
+
+package net.bible.android.view.activity.settings
+
+import android.app.Activity
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
+import androidx.preference.PreferenceDataStore
+import androidx.preference.PreferenceFragmentCompat
+import net.bible.android.activity.R
+import net.bible.android.control.page.window.Window
+import net.bible.android.control.page.window.WindowControl
+import net.bible.android.control.page.window.WindowRepository
+import net.bible.android.database.WorkspaceEntities
+import net.bible.android.view.activity.ActivityScope
+import net.bible.android.view.activity.base.ActivityBase
+import net.bible.android.view.activity.base.CurrentActivityHolder
+import net.bible.android.view.util.locale.LocaleHelper
+import javax.inject.Inject
+
+class ColorSettingsDataStore(val activity: ColorSettingsActivity, val window: Window?, val windowRepository: WindowRepository): PreferenceDataStore() {
+    override fun putInt(key: String?, value: Int) {
+        activity.setDirty()
+        val settings = window?.pageManager?.actualTextDisplaySettings ?: windowRepository.textDisplaySettings
+        val colors = settings.colors?.copy() ?: WorkspaceEntities.TextDisplaySettings.default.colors!!
+
+        when(key) {
+            "text_color_day" -> colors.dayTextColor = value
+            "text_color_night" -> colors.nightTextColor = value
+            "background_color_day" -> colors.dayBackground = value
+            "background_color_night" -> colors.nightBackground = value
+            "noise_day" -> colors.dayNoise = value
+            "noise_night" -> colors.nightNoise = value
+        }
+        if(window != null) {
+            window.pageManager.textDisplaySettings.colors = colors
+        } else {
+            windowRepository.textDisplaySettings.colors = colors
+            windowRepository.updateWindowTextDisplaySettingsValues(WorkspaceEntities.TextDisplaySettings.Types.COLORS, colors)
+        }
+
+    }
+
+    override fun getInt(key: String?, defValue: Int): Int {
+        val colors =
+            if(window != null) {
+                window.pageManager.actualTextDisplaySettings.colors!!
+            } else windowRepository.textDisplaySettings.colors ?: WorkspaceEntities.TextDisplaySettings.default.colors!!
+
+        return when(key) {
+            "text_color_day" -> colors.dayTextColor?: defValue
+            "text_color_night" -> colors.nightTextColor?: defValue
+            "background_color_day" -> colors.dayBackground?: defValue
+            "background_color_night" -> colors.nightBackground?: defValue
+            "noise_day" -> colors.dayNoise?: defValue
+            "noise_night" -> colors.nightNoise?: defValue
+            else -> defValue
+        }
+    }
+}
+
+@ActivityScope
+class ColorSettingsActivity: ActivityBase() {
+    var window: Window? = null
+    private var dirty = false
+
+    @Inject
+    lateinit var windowControl: WindowControl
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.settings_activity)
+        super.buildActivityComponent().inject(this)
+        dirty = false
+
+        CurrentActivityHolder.getInstance().currentActivity = this
+
+        val windowId = intent.extras?.getLong("windowId")
+        window = windowControl.windowRepository.getWindow(windowId)
+
+        supportFragmentManager
+            .beginTransaction()
+            .replace(R.id.settings_container, ColorSettingsFragment(this, window, windowControl.windowRepository))
+            .commit()
+
+        setResult()
+    }
+
+    fun setDirty() {
+        dirty = true
+        setResult()
+    }
+
+    fun setResult() {
+        val resultIntent = Intent(this, ColorSettingsActivity::class.java)
+        val window = this.window
+
+        if(window != null) {
+            title = getString(R.string.window_color_settings_title)
+        } else {
+            title = getString(R.string.workspace_color_settings_title)
+        }
+        resultIntent.putExtra("windowId", window?.id ?: 0L)
+        resultIntent.putExtra("edited", dirty)
+        setResult(Activity.RESULT_OK, resultIntent)
+    }
+
+    override fun attachBaseContext(newBase: Context) {
+        super.attachBaseContext(LocaleHelper.onAttach(newBase))
+    }
+
+    override fun onStop() {
+        super.onStop()
+        Log.i(localClassName, "onStop")
+        // call this onStop, although it is not guaranteed to be called, to ensure an overlap between dereg and reg of current activity, otherwise AppToBackground is fired mistakenly
+        CurrentActivityHolder.getInstance().iAmNoLongerCurrent(this)
+    }
+}
+
+
+class ColorSettingsFragment(val activity: ColorSettingsActivity, val window: Window?, val windowRepository: WindowRepository) : PreferenceFragmentCompat() {
+    override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
+        preferenceManager.preferenceDataStore = ColorSettingsDataStore(activity, window, windowRepository)
+        setPreferencesFromResource(R.xml.color_settings, rootKey)
+    }
+}
